// src/lib/api/generated/models/newEventSubscription.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.0.3
 */
import type { EventSinkID } from './eventSinkID';
import type { EventSubscriptionFilter } from './eventSubscriptionFilter';

export interface NewEventSubscription {
  actions: string[];
  description?: string;
  enabled?: boolean;
  entity_types: string[];
  filter?: EventSubscriptionFilter;
  name: string;
  routing?: unknown;
  sink_id: EventSinkID;
}